Add DLX matrix visualization to SVG export
- improve DLX matrix creation
- Change many int8_t to int32_t

+++ src/sudoku/Setter.h
@@ -8,18 +8,18 @@ class Setter {
 public:
   Setter() = delete;
 
-  static std::unique_ptr<Board> generate(int8_t totalDigits,
+  static std::unique_ptr<Board> generate(int32_t totalDigits,
                                          SymmetryType symmetryType,
                                          const std::vector<std::unique_ptr<AbstractConstraint>>& constraints);
 
 private:
-  static inline bool isOnMainDiagonal(int8_t rowIndex, int8_t columnIndex);
+  static inline bool isOnMainDiagonal(int32_t rowIndex, int32_t columnIndex);
 
-  static inline bool isOnCenter(int8_t rowIndex, int8_t columnIndex);
+  static inline bool isOnCenter(int32_t rowIndex, int32_t columnIndex);
 
-  static std::vector<std::vector<bool>> randomMask(int8_t totalDigits);
+  static std::vector<std::vector<bool>> randomMask(int32_t totalDigits);
 
-  static std::vector<std::vector<bool>> diagonalMirrorMask(int8_t totalDigits);
+  static std::vector<std::vector<bool>> diagonalMirrorMask(int32_t totalDigits);
 
-  static std::vector<std::vector<bool>> diagonalRotationMask(int8_t totalDigits);
+  static std::vector<std::vector<bool>> diagonalRotationMask(int32_t totalDigits);
 };
