+++ src/sudoku/Validator.h
@@ -11,8 +11,8 @@ public:
   static bool checkSolution(const std::vector<std::vector<Sudo>>& board,
                             const std::vector<std::unique_ptr<AbstractConstraint>>& constraints);
 
-  static bool validateMove(const int8_t rowIndex,
-                           const int8_t columnIndex,
+  static bool validateMove(const int32_t rowIndex,
+                           const int32_t columnIndex,
                            const Sudo digit,
                            const std::vector<std::vector<Sudo>>& board,
                            const std::vector<std::unique_ptr<AbstractConstraint>>& constraints);
