+++ src/sudoku/constraints/SudokuBox.h
@@ -15,16 +15,16 @@ public:
   virtual std::string getSvgGroup() const override;
 
   virtual bool validatePlacement(const Sudo digit,
-                                 const int8_t rowIndex,
-                                 const int8_t columnIndex,
+                                 const int32_t rowIndex,
+                                 const int32_t columnIndex,
                                  const std::vector<std::vector<Sudo>> board) const override;
 
   virtual bool satisfy(std::vector<std::vector<Sudo>> board) const override;
 
   virtual int32_t getDLXConstraintColumnsAmount() const override;
 
-  virtual bool getDLXConstraint(const int8_t boardRow,
-                                const int8_t boardColumn,
+  virtual bool getDLXConstraint(const int32_t boardRow,
+                                const int32_t boardColumn,
                                 const int32_t columnId,
                                 const Sudo possibleDigit,
                                 const Sudo actualDigit) const override;
@@ -36,12 +36,12 @@ private:
    * @param columnIndex A cell's column index in the grid
    * @return The "box" index for that cell
    */
-  static int8_t getBoxIndex(const int8_t rowIndex, const int8_t columnIndex);
+  static int32_t getBoxIndex(const int32_t rowIndex, const int32_t columnIndex);
 
   /**
    * Creates all [row][column] index pairs for the cells of every box
    * @return A vector of vectors of index pairs. Outer vector references each box, inner vector contains all index
    * pairs for that box
    */
-  static std::vector<std::vector<std::pair<int8_t, int8_t>>> getBoxIndices();
+  static std::vector<std::vector<std::pair<int32_t, int32_t>>> getBoxIndices();
 };
