+++ src/sudoku/constraints/SudokuCell.h
@@ -15,16 +15,16 @@ public:
   virtual std::string getSvgGroup() const override;
 
   virtual bool validatePlacement(const Sudo digit,
-                                 const int8_t rowIndex,
-                                 const int8_t columnIndex,
+                                 const int32_t rowIndex,
+                                 const int32_t columnIndex,
                                  const std::vector<std::vector<Sudo>> board) const override;
 
   virtual bool satisfy(std::vector<std::vector<Sudo>> board) const override;
 
   virtual int32_t getDLXConstraintColumnsAmount() const override;
 
-  virtual bool getDLXConstraint(const int8_t boardRow,
-                                const int8_t boardColumn,
+  virtual bool getDLXConstraint(const int32_t boardRow,
+                                const int32_t boardColumn,
                                 const int32_t columnId,
                                 const Sudo possibleDigit,
                                 const Sudo actualDigit) const override;
